Move stuff out of the root of stock-common into an all/ subdir.
Thus it is nolonger necessary to checkout stock-common (without subdirs) to get the primary build file.

The individual project build files are configured to work with a directory structure like:
   stock-common         : not from svn
   stock-common/all     : svn://stocksoftware/ss/stock-common/all
   stock-common/iface   : svn://stocksoftware/ss/stock-common/iface/trunk
   stock-common/message : svn://stocksoftware/ss/stock-common/message/trunk
   stock-common/ui      : svn://stocksoftware/ss/stock-common/ui/trunk
   etc

If you wish to alter this then see comments in 'all/project.properties' and the individual project 'project.properties' files to get it all to work.